CFBD Venues Endpoint Overview

Pulls all college football venues and data on capacity, grass, city/state, location, elevation, dome, timezone and construction year.

Usage

cfbd_venues()

Value

A data frame with 337 rows and 13 variables:

venue_id:integer.

Referencing venue ID.

name:character.

Venue name.

capacity:integer.

Stadium capacity.

grass:logical.

TRUE/FALSE response on whether the field is grass or not (oh, and there are so many others).

city:character.

Venue city.

state:character.

Venue state.

zip:character.

Venue zip.

country_code:character.

Venue country code.

location:list.

Venue location.

elevation:character.

Venue elevation.

year_constructed:integer.

Year in which the venue was constructed.

dome:logical.

TRUE/FALSE response to whether the venue has a dome or not.

timezone:character.

Time zone in which the venue resides (i.e. Eastern Time -> "America/New York").

Details

CFB Venue Information

cfbd_venues()

Examples

# \donttest{
  try(cfbd_venues())
#> ── Venue data from CollegeFootballData.com ─────────────────── recruitR 0.0.3 ──
#> ℹ Data updated: 2022-06-17 20:42:49 UTC
#> # A tibble: 351 × 13
#>    venue_id name        capacity grass city  state zip   country_code location$x
#>       <int> <chr>          <int> <lgl> <chr> <chr> <chr> <chr>             <dbl>
#>  1     4747 Abbott Mem…    10000 NA    Tusk… AL     NA   US                 32.4
#>  2     3599 Ace W. Mum…    28500 FALSE Bato… LA    "708… US                 30.5
#>  3     3601 Aggie Memo…    28853 FALSE Las … NM    "880… US                 32.3
#>  4     3602 Aggie Stad…    10743 FALSE Davis CA    "956… US                 38.5
#>  5     3603 Aggie Stad…    21500 TRUE  Gree… NC    "274… US                 36.1
#>  6     3604 Alamodome      36582 FALSE San … TX    "782… US                 29.4
#>  7     3653 Albertsons…    36387 FALSE Boise ID    "837… US                 43.6
#>  8     3605 Alerus Cen…    12283 FALSE Gran… ND    "582… US                 47.9
#>  9     1027 Alex G. Sp…    11075 TRUE  San … CA    "934… US                 35.3
#> 10     3607 Alfond Sta…    10000 FALSE Orono ME    "446… US                 44.9
#> # … with 341 more rows, and 5 more variables: location$y <dbl>,
#> #   elevation <chr>, year_constructed <int>, dome <lgl>, timezone <chr>
# }