Get 247 Recruitment

Get 247 Recruitment

Usage

tfs_recruitment(recruitment_id)

Arguments

recruitment_id

(Integer)

Value

tfs_recruitment()

Examples

# \donttest{
   tfs_recruitment(recruitment_id = 130361)
#>      key player_key player_hometown_state player_hometown_city
#> 1 130361   46084728               Georgia              Suwanee
#>   player_first_name player_last_name player_full_name player_height
#> 1            Travis           Hunter    Travis Hunter         6-0.5
#>   player_weight
#> 1           188
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             player_bio
#> 1 Member of 247Sports’ 2022 Freaks List.\r\nAlso plays basketball.\r\n\r\n2021: Two-way player for a Collins Hill squad that captured a Peach State AAAAAAA title. Missed five games due to an ankle injury. Still caught 76 passes for 1,128 and 10 touchdowns in nine games. Credited with 20.5 tackles, 4 interceptions and a forced fumble.\r\n\r\n2020: Named the MaxPreps Junior of the Year; MaxPreps Junior All-American. Best player on a Collins Hill team that went 12-3 and made AAAAAAA state final. Got snaps at both cornerback and wide receiver. Caught 137 passes for 1,731 yards and 24 touchdowns.\r\n\r\n2019: All-county selection by the Gwinnett Daily Post and the Touchdown Club of Gwinnett. Helped Collins Hill to a 6-5 record and Georgia AAAAAAA first-round playoff appearance as a sophomore. Two-way player that also totaled 919 receiving yards and 12 touchdowns.
#>        player_birthdate player_modified_user player_modified_date
#> 1 5/18/2003 12:00:00 AM         ChanceLinton 8/30/2025 4:19:53 PM
#>   player_cbs_key                                           player_url
#> 1       28909849 https://247sports.com/player/travis-hunter-46084728/
#>   player_player_high_school_name player_last_recruitment_player_institution
#> 1                   Collins Hill                                     236028
#>   player_current_player_institution player_twitter_contact
#> 1                            344052                  47823
#>   player_primary_player_sport player_primary_player_position_abbreviation
#> 1                      354247                                         ATH
#>   player_default_name player_default_asset
#> 1       Travis Hunter             12616559
#>                                            player_default_asset_url
#> 1 https://s3media.247sports.com/Uploads/Assets/559/616/12616559.jpg
#>   player_pro_stat_player player_college_stat_player
#> 1                 150695                     150695
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  player_bio_or_default
#> 1 Member of 247Sports’ 2022 Freaks List.\r\nAlso plays basketball.\r\n\r\n2021: Two-way player for a Collins Hill squad that captured a Peach State AAAAAAA title. Missed five games due to an ankle injury. Still caught 76 passes for 1,128 and 10 touchdowns in nine games. Credited with 20.5 tackles, 4 interceptions and a forced fumble.\r\n\r\n2020: Named the MaxPreps Junior of the Year; MaxPreps Junior All-American. Best player on a Collins Hill team that went 12-3 and made AAAAAAA state final. Got snaps at both cornerback and wide receiver. Caught 137 passes for 1,731 yards and 24 touchdowns.\r\n\r\n2019: All-county selection by the Gwinnett Daily Post and the Touchdown Club of Gwinnett. Helped Collins Hill to a 6-5 record and Georgia AAAAAAA first-round playoff appearance as a sophomore. Two-way player that also totaled 919 receiving yards and 12 touchdowns.
#>   player_rating player_star_rating player_national_rank player_position_rank
#> 1           100                  5                    1                    1
#>   player_state_rank player_institution year signed_institution position
#> 1                 1             236028 2022              24245       24
#>   institution state player_sport composite_strength final_choice
#> 1        9675     3       246045                  3       295497
#>   highest_recruit_interest_event_type highest_recruit_interest_event
#> 1                            Enrolled                         358456
#>   committed_recruit_interest committed_institution highest_recruit_interest
#> 1                     761045                 24245                   761045
#>   primary_player_position primary_position  default_name
#> 1                  273218               24 Travis Hunter
#>                                     commited_institution_team_image
#> 1 https://s3media.247sports.com/Uploads/Assets/845/542/12542845.png
#>   recruit_interest_count
#> 1                     17
#>                                                      recruit_interests_url
#> 1 https://247sports.com/recruitment/travis-hunter-130361/recruitinterests/
# }