Get 247 Recruitment

Get 247 Recruitment

tfs_recruitment(recruitment_id)

Arguments

recruitment_id

(Integer)

Value

tfs_recruitment()

Examples

# \donttest{
   tfs_recruitment(recruitment_id = 130361)
#>      key player_key player_hometown_state player_hometown_city
#> 1 130361   46084728               Georgia              Suwanee
#>   player_first_name player_last_name player_full_name player_height
#> 1            Travis           Hunter    Travis Hunter           6-1
#>   player_weight
#> 1           165
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   player_bio
#> 1 Also plays basketball.\r\n\r\n2021: Two-way player for a Collins Hill squad that captured a Peach State AAAAAAA title. Missed five games due to an ankle injury. Still caught 76 passes for 1,128 and 10 touchdowns in nine games. Credited with 20.5 tackles, 4 interceptions and a forced fumble.\r\n\r\n2020: Named the MaxPreps Junior of the Year; MaxPreps Junior All-American. Best player on a Collins Hill team that went 12-3 and made AAAAAAA state final. Got snaps at both cornerback and wide receiver. Caught 137 passes for 1,731 yards and 24 touchdowns.\r\n\r\n2019: All-county selection by the Gwinnett Daily Post and the Touchdown Club of Gwinnett. Helped Collins Hill to a 6-5 record and Georgia AAAAAAA first-round playoff appearance as a sophomore. Two-way player that also totaled 919 receiving yards and 12 touchdowns.
#>   player_modified_user player_modified_date player_cbs_key
#> 1          Chad Carson 6/2/2022 10:48:46 AM       26902531
#>                                             player_url
#> 1 https://247sports.com/Player/Travis-Hunter-46084728/
#>   player_player_high_school_name player_last_recruitment_player_institution
#> 1                   Collins Hill                                     236028
#>   player_current_player_institution player_twitter_contact
#> 1                            236028                  47823
#>   player_primary_player_sport player_primary_player_position_abbreviation
#> 1                      246045                                          CB
#>   player_primary_recruitment player_default_name player_default_asset
#> 1                     130361       Travis Hunter             10864998
#>                                            player_default_asset_url
#> 1 https://s3media.247sports.com/Uploads/Assets/998/864/10864998.JPG
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        player_bio_or_default
#> 1 Also plays basketball.\r\n\r\n2021: Two-way player for a Collins Hill squad that captured a Peach State AAAAAAA title. Missed five games due to an ankle injury. Still caught 76 passes for 1,128 and 10 touchdowns in nine games. Credited with 20.5 tackles, 4 interceptions and a forced fumble.\r\n\r\n2020: Named the MaxPreps Junior of the Year; MaxPreps Junior All-American. Best player on a Collins Hill team that went 12-3 and made AAAAAAA state final. Got snaps at both cornerback and wide receiver. Caught 137 passes for 1,731 yards and 24 touchdowns.\r\n\r\n2019: All-county selection by the Gwinnett Daily Post and the Touchdown Club of Gwinnett. Helped Collins Hill to a 6-5 record and Georgia AAAAAAA first-round playoff appearance as a sophomore. Two-way player that also totaled 919 receiving yards and 12 touchdowns.
#>   player_composite_rating player_composite_star_rating player_national_rank
#> 1        99.9888517279822                            5                    1
#>   player_position_rank player_state_rank player_institution year
#> 1                    1                 1             236028 2022
#>   signed_institution position institution state player_sport composite_strength
#> 1              24245       24        9675     3       246045                  3
#>   highest_recruit_interest_event_type highest_recruit_interest_event
#> 1                              Signed                         348545
#>   committed_recruit_interest committed_institution highest_recruit_interest
#> 1                     761045                 24245                   761045
#>   primary_player_position primary_position  default_name
#> 1                  273218               24 Travis Hunter
#>                                    commited_institution_team_image
#> 1 https://s3media.247sports.com/Uploads/Assets/783/972/9972783.png
#>   recruit_interest_count
#> 1                     17
#>                                                      recruit_interests_url
#> 1 https://247sports.com/Recruitment/Travis-Hunter-130361/RecruitInterests/
# }